verify hardening, enable-backend guards, refactors, docs
