+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog, and this project follows Semantic Versioning.
 
-## [1.1.3] - 2026-02-05
+## [1.2.0] - 2026-02-05
 
 - Add automatic interface assignment for Kea when using `--create-subnets` (populates listening interfaces).
 - Add automatic interface assignment for dnsmasq when using `--create-subnets`.
 - Add `--enable-backend` flag to disable ISC DHCP on migrated interfaces and enable the target backend.
+- Make ISC DHCP disablement compatible with OPNsense by removing `<enable>` tags instead of writing `0`.
+- Guard `--enable-backend` so it refuses to run if ISC DHCP already appears disabled (prevents dual-DHCP on repeat runs).
+- Verify now normalizes XML before diffing, avoiding noisy formatting-only diffs.
+- Add post-convert verification that ISC DHCP disablement actually took effect.
+- Refactor CLI and migrate modules for maintainability.
+- Update README scripted usage, add CI/release badges, and document enable-backend workflow notes.
 
 ## [1.1.2] - 2026-02-04
