+++ src/migrate/kea/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) use convert::convert_kea;
+pub(crate) use scan::scan_kea;
+
+mod convert;
+mod scan;
